исправление замечаний: исправление sql запросов в  findUserWithCommonLikes() и в getFilmRecommendation(), перенос getRecommendations() в FilmService
