+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -2,13 +2,13 @@ package ru.yandex.practicum.filmorate.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
+import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
 import java.time.LocalDate;
 import java.util.Collection;
@@ -27,6 +27,7 @@ public class FilmService {
 
     private final FilmStorage filmStorage;
     private final UserService userService;
+    private final UserStorage userStorage;
     private final FeedService feedService;
     private final DirectorService directorService;
 
@@ -147,5 +148,12 @@ public class FilmService {
 
     public Collection<Film> getSearchResults(String query, List<String> by) {
         return filmStorage.getSearchResults(query, by);
+    }
+
+    public Collection<Film> getRecommendations(int userWantsRecomId) {
+        log.info("Найден пользователь с похожими лайками");
+        int userWithCommonLikesId = userStorage.findUserWithCommonLikes(userWantsRecomId);
+        log.info("Выгружаем список рекомендованных фильмов для пользователя {}", userWantsRecomId);
+        return filmStorage.getFilmRecommendation(userWantsRecomId, userWithCommonLikesId);
     }
 }
