+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -2,12 +2,10 @@ package ru.yandex.practicum.filmorate.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Feed;
-import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
 import ru.yandex.practicum.filmorate.storage.user.UserStorage;
@@ -113,10 +111,4 @@ public class UserService {
         userStorage.deleteById(userId);
     }
 
-    public Collection<Film> getRecommendations(int userWantsRecomId) {
-        log.info("Найден пользователь с похожими лайками");
-        int userWithCommonLikesId = userStorage.findUserWithCommonLikes(userWantsRecomId);
-        log.info("Выгружаем список рекомендованных фильмов для пользователя {}", userWantsRecomId);
-        return filmStorage.getFilmRecommendation(userWantsRecomId, userWithCommonLikesId);
-    }
 }
