+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -26,7 +26,4 @@ public interface UserStorage {
     void deleteById(int userId);
 
     Integer findUserWithCommonLikes(int userId);
-
-    Integer findUserWithCommonLikes(int userId);
-
 }
