+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/impl/UserDbStorage.java
@@ -2,7 +2,6 @@ package ru.yandex.practicum.filmorate.storage.user.impl;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
@@ -145,9 +144,10 @@ public class UserDbStorage implements UserStorage {
     public Integer findUserWithCommonLikes (int userWantsRecomId) {
         String sqlQuery = "SELECT fl2.user_id " +
                 "FROM FILM_LIKES AS fl1, FILM_LIKES AS fl2 " +
-                "WHERE fl1.USER_ID = ? AND fl1.USER_ID != fl2.USER_ID " +
+                "WHERE fl1.film_id = fl2.film_id " +
+                "AND fl1.USER_ID = ? AND fl1.USER_ID != fl2.USER_ID " +
                 "GROUP BY fl1.user_id, fl2.user_id " +
-                "ORDER BY count(fl2.USER_ID) desc limit 1";
+                "ORDER BY count(*) desc limit 1";
         try {
             return jdbcTemplate.queryForObject(sqlQuery, Integer.class, userWantsRecomId);
         } catch (EmptyResultDataAccessException exception) {
